back/requirements.txt
Django==2.1.3
django-braces==1.13.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
gunicorn==19.9.0
Pillow==5.3.0
psycopg2-binary==2.7.6.1
pytz==2018.7
six==1.12.0
